refactor(): RacingCar 클래스명 변경 및 controller 패키지 생성
